Added the following features:
as-osadecompile-command: defcustom, set to osadecompile by default.

as-execute-line: sends the current line to applescript.
load-and-run-applescript: will load and run the applescript at
``path'', optionally using the alist ``tokens'' to make replacements
before sending.

as-quote: Quotes a string for use in an applescript statement.

Corrected syntax entry for backslash.

Added the ability to read and write compiled
applescripts (decompile-scpt and compile-scpt).

Added .scpt to auto-mode-alist

Added a shortcut (C-c RET) for inserting applescript linebreak character ("¬").